Refactor scoring into dedicated module
