+++ src/searchCore.ts
@@ -2,7 +2,8 @@ import { stemmer } from 'stemmer';
 import { annotateAvailability } from './availability';
 import { generateCandidates } from './generator';
 import { getPerformance, type TimedPerformance } from './perf';
-import { rankDomains, rerankWithUnavailable, scoreCandidates } from './ranking';
+import { rankDomains, rerankWithUnavailable } from './ranking';
+import { scoreCandidates } from './scoring';
 import { expandSynonyms } from './synonyms';
 import { ClientInitOptions, DomainCandidate, DomainSearchOptions, LatencyMetrics, ProcessedQuery, SearchResponse } from './types';
 import { getCcTld, isValidTld, normalizeTld, normalizeTokens, unique } from './utils';
